% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadia-package.R
\docType{package}
\name{circadia-package}
\alias{circadia}
\alias{circadia-package}
\title{circadia: Circadia Lab Colour Palettes and ggplot2 Theme}
\description{
Provides the shared visual identity for the Circadia Lab R ecosystem: colour palettes, ggplot2 theme, and ggplot2 scales. Palettes cover qualitative, sequential, and diverging use cases anchored on the lab's brand colours. Designed to be used alongside zeitR, slumbR, tallieR, and syncR.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/circadia-bio/circadia}
  \item Report bugs at \url{https://github.com/circadia-bio/circadia/issues}
}

}
\author{
\strong{Maintainer}: Lucas França \email{lucas.franca@northumbria.ac.uk} (\href{https://orcid.org/0000-0003-0853-1319}{ORCID})

Authors:
\itemize{
  \item Lucas França \email{lucas.franca@northumbria.ac.uk} (\href{https://orcid.org/0000-0003-0853-1319}{ORCID})
  \item Mario Leocadio-Miguel \email{mario.miguel@northumbria.ac.uk} (\href{https://orcid.org/0000-0002-7248-3529}{ORCID})
}

}
\keyword{internal}
