% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{circadia_palettes}
\alias{circadia_palettes}
\title{List all available Circadia Lab palettes}
\usage{
circadia_palettes()
}
\value{
A named integer vector of palette lengths, invisibly.
}
\description{
Prints the names and sizes of all palettes defined in the package.
}
\examples{
circadia_palettes()

}
