% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{.circadia_palettes}
\alias{.circadia_palettes}
\title{Circadia Lab colour palettes}
\format{
A named list of character vectors, each containing hex colour codes.
}
\usage{
.circadia_palettes
}
\description{
Named list of all palettes available in the circadia package.
}
\keyword{internal}
