% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_circadia}
\alias{scale_colour_circadia}
\alias{scale_color_circadia}
\title{Circadia Lab discrete colour scale}
\usage{
scale_colour_circadia(palette = "main", reverse = FALSE, ...)

scale_color_circadia(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette name. Default \code{"main"}.}

\item{reverse}{Logical. Reverse the palette? Default \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
Applies a Circadia Lab qualitative palette to the \code{colour} aesthetic.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg, colour = factor(cyl))) +
  geom_point(size = 3) +
  scale_colour_circadia() +
  theme_circadia()

}
