% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_circadia}
\alias{scale_fill_circadia}
\title{Circadia Lab discrete fill scale}
\usage{
scale_fill_circadia(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette name. Default \code{"main"}.}

\item{reverse}{Logical. Reverse the palette? Default \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
Applies a Circadia Lab qualitative palette to the \code{fill} aesthetic.
}
\examples{
library(ggplot2)
ggplot(mpg, aes(class, fill = drv)) +
  geom_bar() +
  scale_fill_circadia() +
  theme_circadia()

}
