% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_circadia_c}
\alias{scale_fill_circadia_c}
\title{Circadia Lab continuous fill scale}
\usage{
scale_fill_circadia_c(palette = "blues", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette name. Default \code{"blues"}.}

\item{reverse}{Logical. Reverse the palette? Default \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_colour_gradientn]{ggplot2::scale_colour_gradientn()}}.}
}
\description{
Interpolates across a Circadia Lab palette for continuous \code{fill} data.
}
