% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{.circadia_domain_info}
\alias{.circadia_domain_info}
\title{Domain colour map}
\format{
An object of class \code{list} of length 8.
}
\usage{
.circadia_domain_info
}
\description{
Associates each data domain in the Circadia Lab ecosystem with a brand
colour. Consumed internally by \code{\link[=domain_colour_for]{domain_colour_for()}}.
}
\keyword{internal}
